configure: remove AG_GST_PARSE_SUBSYSTEM_DISABLES
This would check which subsystems are disabled in core by grepping
gstconfig.h. Only problem is: gstconfig.h has moved into libdir now
so we've been checking a non-existent file for a while now. The
macro would just sets GST_DISABLE_* for use in configure.ac and
Makefile.am, but we don't use that anywhere so just get rid of it
(the one place where we use GST_DISABLE_GST_DEBUG is in a .c file
which gets the define from the gstconfig.h include).

https://bugzilla.gnome.org/show_bug.cgi?id=750056